# config/swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "1.0"
  title: "COFFIT API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "coffit@swm.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "127.0.0.1:3000"
basePath: "/"
...
schemes:
  - "https"
  - "http"
components:
  schemas:
    PTwithScheduleAndComment:
      type: "object"
      properties:
        id: #PK
          type: "integer"
          format: "int64"
          description: "PK"
        state:
          type: "string"
          description: "PT state"
          enum:
            - "exercising"
            - "finished"
        total_number:
          type: "integer"
          format: "int32"
        rest_number:
          type: "integer"
          format: "int32"
        start_date:
          type: "integer"
          format: "date"
          description: "first start PT date"
        end_date:
          type: "integer"
          format: "date"
          description: "Termination time"
        student_id: #FK
          type: "integer"
          format: "int64"
          description: "FK"
        trainer_id: #FK
          type: "integer"
          format: "int64"
          description: "FK"
        PTComment:
          $ref: "#/definitions/PTComment"
        schedules:
          type: array
          items:
            $ref: "#/definitions/Schedule"
...
    TrainerDetailWithPictures:
      type: object
      properties:
        id:
          type: "integer"
          format: "int64"
        username:
          type: "string"
        price:
          type: "integer"
          format: "int64"
        career:
          type: "string"
        description:
          type: "string"
        summary:
          type: "string"
        picture_url:
          type: "string"
        phone_number:
          type: "string"
        fcm_token:
          type: "string"
        total_star:
          type: "integer"
        num_review:
          type: "integer"
        created_at:
          type: integer
          format: date
        updated_at:
          type: integer
          format: date
        trainer_accditional_picture:
          type: array
          items:
            $ref: "#/definitions/TrainerPicture"
    StudentWithSchedules:
      type: object
      properties:
        id:
          type: "integer"
          format: "int64"
        uername:
          type: "string"
        email:
          type: "string"
        age:
          type: "integer"
          format: "int64"
        picture:
          type: "string"
        gender:
          type: "string"
        phone_number:
          type: string
          pattern: '^\d{3}-\d{4}-\d{4}$'
        fcm_token:
          type: "string"
        created_at:
          type: integer
          format: date
        updated_at:
          type: integer
          format: date
        schedules:
          type: array
          items:
            $ref: "#/definitions/Schedule"
paths:
...
  /trainers/{Id}: # trainer picture 추가하기
    get:
      tags:
        - "trainer"
      summary: "Find trainer by ID"
      description: "Returns trainer description"
      operationId: "getPetById"
      produces:
        - "application/json"
      parameters:
        - name: "Id"
          in: "path"
          description: "ID of trainer"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/components/schemas/TrainerDetailWithPictures"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Trainer not found"
    put:
      tags:
        - "trainer"
      summary: "Updates a trainer information"
      description: ""
      operationId: "updateTrainerWithForm"
      consumes:
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - name: "Id"
          in: "path"
          description: "ID of pet that needs to be updated"
          required: true
          type: "integer"
          format: "int64"
        - name: "body"
          in: "body"
          schema:
            $ref: "#/components/schemas/TrainerDetailWithPictures"
      responses:
        204:
          description: "your request applied"
        405:
          description: "Invalid input"
#  /trainers/{Id}/profilePicture:
#    post:
#      tags:
#        - "trainer"
#      summary: "Already exist file remove and update. but, when not exist profile picture, upload picture"
#      description: "Already exist file remove and update. but, when not exist profile picture, upload picture"
#      operationId: "uploadFile"
#      consumes:
#        - "multipart/form-data"
#      produces:
#        - "application/json"
#      parameters:
#        - name: "Id"
#          in: "path"
#          description: "ID of trainer to update"
#          required: true
#          type: "integer"
#          format: "int64"
#        - name: "file"
#          in: "formData"
#          description: "file to upload"
#          required: false
#          type: "file"
#      responses:
#        200:
#          description: "successful operation"
#  /trainers/{Id}/subPictures:
#    post:
#      tags:
#        - "trainer"
#      summary: "This is activity picture collorection. If already exist file remove and update. but, when not exist profile picture, upload picture"
#      description: ""
#      operationId: "uploadSubPicture"
#      consumes:
#        - "multipart/form-data"
#      produces:
#        - "application/json"
#      parameters:
#        - name: "Id"
#          in: "path"
#          description: "ID of trainer to update"
#          required: true
#          type: "integer"
#          format: "int64"
#        - name: "pictures"
#          in: "formData"
#          description: "array of file to upload"
#          required: true
#          type: array
#          items:
#            type: string
#      responses:
#        200:
#          description: "successful operation"
  /trainers/{Id}/token:
    post:
      tags:
        - "trainer"
      summary: "upload Token"
      description: ""
      operationId: "setToken"
      produces:
        - "application/json"
      parameters:
        - name: "Id"
          in: "path"
          description: "ID of trainer"
          required: true
          type: "integer"
          format: "int64"
        - name: "token"
          in: "path"
          description: "token of trainer"
          required: true
          type: "string"
      responses:
        201:
          description: "token saved successfully!!"
        405:
          description: "Invalid input"
...
  /students/trainers/{id}: # 트레이너가 가지고 있는 학생 보여주기
    get:
      tags:
        - "student"
      summary: "Get all student including certain trainer"
      description: "Get student information using trainer id"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: integer
          format: int64
        - name: "studentName"
          in: "query"
          required: false
          type: string
          description: "If you want to search student name, you can find student information including schedule."
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/components/schemas/StudentWithSchedules"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Not found"
...
  /PTs/trainers/{id}:
    get:
      tags:
        - "PT"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of trainer"
          required: true
          type: "integer"
          format: "int64"
      summary: "get PT list from trainer and trainer today schedule"
      description: "Get all PT, Nessary student id using session or JWT"
      operationId: "get PT list from trainer"
      produces:
        - "application/json"
      responses:
        200:
          description: "Success!!"
          schema:
            $ref: "#/components/schemas/PTwithScheduleAndComment"
...
  /notifications/trainers/{id}:
    get:
      tags:
        - "notification"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of trainer"
          required: true
          type: "integer"
          format: "int64"
      summary: "get certain trainer notification"
      description: ""
      produces:
        - "application/json"
      responses:
        200:
          description: "Success!!"
          schema:
            type: array
            items:
              $ref: "#/definitions/Notification"
...
  /schedules/trainer/{id}:
    get:
      tags:
        - "schedule"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of trainer"
          required: true
          type: "integer"
          format: "int64"
      summary: "get certain trainer schedules "
      responses:
        200:
          description: ""
          schema:
            type: object
            properties:
              trainer:
                $ref: '#/definitions/Trainer'
              schedules:
                type: array
                items:
                  $ref: '#/definitions/Schedule'
              availableTime:
                type: array
                items:
                  $ref: '#/definitions/TrainerSchedule'
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Trainer:
    type : "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      price:
        type: "integer"
        format: "int64"
      career:
        type: "string"
      description:
        type: "string"
      summary:
        type: "string"
      picture_url:
        type: "string"
      phone_number:
        type: "string"
      fcm_token:
        type: "string"
      total_star:
        type: "integer"
      num_review:
        type: "integer"
      created_at:
        type: integer
        format: date
      updated_at:
        type: integer
        format: date
  TrainerPicture:
    type : "object"
    properties:
      id: #PK
        type: "integer"
        format: "int64"
      picture_url:
        type: "string"
      trainer_id: #FK
        type: "integer"
        format: "int64"
      created_at:
        type: integer
        format: date
      updated_at:
        type: integer
        format: date
  TrainerSchedule:
    type: object
    properties:
      id: #PK
        type: "integer"
        format: "int64"
      start_time:
        type: integer
        format: date
      availability :
        type: boolean
        default: true
      created_at:
        type: integer
        format: date
      updated_at:
        type: integer
        format: date
      trainer_id: #FK
        type: "integer"
        format: "int64"
...
  Schedule:
    type : "object"
    properties:
      id: #PK
        type: "integer"
        format: "int64"
      state:
        type: "integer"
        format: "int64"
      date:
        type: "string"
        format: "date"
      start_time:
        type: "integer"
        format: "time"
      end_time:
        type: "integer"
        format: "time"
      memo:
        type: "string"
      created_at:
        type: "integer"
        format: "date"
      updated_at:
        type: "integer"
        format: "date"
      student_id:
        type: "integer"
        format: "int64"
      trainer_id:
        type: "integer"
        format: "int64"
      PT_id: #FK
        type: "integer"
        format: "int64"
...
  Notification:
    type: object
    properties:
      id: #PK
        type: integer
        format: int64
      towhom:
        type: integer
        format: int32
        enum:
          - 0 #Sending trainer
          - 1 #Sending student
      contents:
        type: string
      type:
        type: integer
      schedule_id:
        type: integer
        format: int64
      student_id:
        type: integer
        format: int64
      trainer_id:
        type: integer
        format: int64
      origin_date:
        type: integer
        format: date
      request_date:
        type: integer
        format: date
      created_at:
        type: integer
        format: date
      updated_at:
        type: integer
        format: date
  PTComment:
    type: object
    properties:
      id:
        type: integer
        format: int64
      comment:
        type: string
      PT_id:
        type: integer
        format: int64
      created_at:
        type: integer
        format: date
      updated_at:
        type: integer
        format: date
...